feat(server):create default standard blueprints with new workspaces
